Bump versions and revert file move
