+++ app/src/androidTest/java/journal/gratitude/com/gratitudejournal/di/TestCloudUploadModule.kt
@@ -3,7 +3,7 @@ package journal.gratitude.com.gratitudejournal.di
 import dagger.Module
 import dagger.Provides
 import journal.gratitude.com.gratitudejournal.fakes.FakeCloudUploader
-import journal.gratitude.com.gratitudejournal.util.backups.dropbox.CloudProvider
+import journal.gratitude.com.gratitudejournal.util.backups.CloudProvider
 import javax.inject.Singleton
 
 @Module
