+++ app/src/main/java/journal/gratitude/com/gratitudejournal/di/CloudUploadModule.kt
@@ -3,7 +3,7 @@ package journal.gratitude.com.gratitudejournal.di
 import android.content.Context
 import dagger.Module
 import dagger.Provides
-import journal.gratitude.com.gratitudejournal.util.backups.dropbox.CloudProvider
+import journal.gratitude.com.gratitudejournal.util.backups.CloudProvider
 import journal.gratitude.com.gratitudejournal.util.backups.dropbox.DropboxUploader
 import javax.inject.Singleton
