+++ versions.gradle
@@ -1,7 +1,7 @@
 ext {
     major = 1
     minor = 8
-    patch = 1
+    patch = 3
     beta = ""
 
     androidCompileSdk = 29
